this backlog_factor is sent in info-response, make sure we don't include floats in it
git-svn-id: https://xpra.org/svn/Xpra/trunk@17037 3bb7dfac-3a0b-4e04-842a-767bc560f471

+++ src/xpra/server/window/batch_delay_calculator.py
@@ -210,7 +210,7 @@ def get_target_quality(window_dimensions, batch, global_statistics, statistics, 
     packets_backlog, pixels_backlog, _ = statistics.get_client_backlog()
     pb_ratio = pixels_backlog/low_limit
     pixels_bl = 1.0 - logp(pb_ratio//4)     #4 frames behind -> min quality
-    info["backlog_factor"] = packets_backlog, pixels_backlog, low_limit, pb_ratio, int(100.0*pixels_bl)
+    info["backlog_factor"] = packets_backlog, pixels_backlog, low_limit, int(pb_ratio), int(100.0*pixels_bl)
     target = pixels_bl
     if batch is not None:
         recs = len(batch.last_actual_delays)
